Rename InGameInputActionsHandler to InputListener
